Don't force using the inferred parser
The inferred parser does not take into account any overrides specified
in .prettierrc.

By not specifying a parser in the options, prettier shall still infer
the parser internally, at a point where overrides specified in
.prettierrc are resolved.

This means that users can specify parsers for custom file extensions
rather than be limited to the filename->parser mapping that prettier has
built-in.

Fixes #1